requirements.txt
anthropic
requests
beautifulsoup4
jinja2
python-dotenv
